Init the effort of rebuild Mixt framework in ATS
I choose to follow the origin MIXT, which is a name of Media-IXT,
and I don't have any idea what does that IXT mean.

please find out the origin product description on the google, for
example:

http://marc.info/?l=ms-wmtalk&m=103590922614358&w=2
http://www.docin.com/p-34149349.html

all those are very old docs, but as you know, many very old Inktomi
systems still running in the corner of the world.